+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.2.0] 2021-11-02
+### Added
+- Comments mode improved with C-style multi-line comment support
+- Option `--tag` added, to choose a specific starting tag for one-line comments
+### Changed
+- Dictionaries improved using OWASP code review guide and code review
+  repositories
+
 ## [1.1.0] 2021-06-02
 ### Added
 - "Strings" search mode for strings on one line starting and ending with `"`